requirements.txt
# Core Libraries
openai==0.28.0
pyyaml==6.0.2
